Rename table models
